@@ -1,13 +1,14 @@
 module github.com/terraform-google-modules/terraform-google-kubernetes-engine/test/integration
 
-go 1.21
+go 1.22.0
 
-toolchain go1.21.5
+toolchain go1.22.7
 
 require (
-	github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test v0.10.3
-	github.com/gruntwork-io/terratest v0.46.8
-	github.com/stretchr/testify v1.8.4
+	github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test v0.16.2
+	github.com/gruntwork-io/terratest v0.47.1
+	github.com/stretchr/testify v1.9.0
+	github.com/tidwall/gjson v1.17.3
 )
 
 require (
